+++ modules/canvaskit/rt_shader.js
@@ -16,18 +16,22 @@ CanvasKit._extraInitializations.push(function() {
   };
 
   CanvasKit.RuntimeEffect.prototype.makeShader = function(floats, localMatrix) {
-    // We don't need to free these floats because they will become owned by the shader.
-    var fptr = copy1dArray(floats, "HEAPF32");
+    // If the uniforms were set in a MallocObj, we don't want the shader to take ownership of
+    // them (and free the memory when the shader is freed).
+    var shouldOwnUniforms = !floats['_ck'];
+    var fptr = copy1dArray(floats, 'HEAPF32');
     var localMatrixPtr = copy3x3MatrixToWasm(localMatrix);
     // Our array has 4 bytes per float, so be sure to account for that before
     // sending it over the wire.
-    return this._makeShader(fptr, floats.length * 4, localMatrixPtr);
+    return this._makeShader(fptr, floats.length * 4, shouldOwnUniforms, localMatrixPtr);
   }
 
   // childrenWithShaders is an array of other shaders (e.g. Image.makeShader())
   CanvasKit.RuntimeEffect.prototype.makeShaderWithChildren = function(floats, childrenShaders, localMatrix) {
-    // We don't need to free these floats because they will become owned by the shader.
-    var fptr = copy1dArray(floats, "HEAPF32");
+    // If the uniforms were set in a MallocObj, we don't want the shader to take ownership of
+    // them (and free the memory when the shader is freed).
+    var shouldOwnUniforms = !floats['_ck'];
+    var fptr = copy1dArray(floats, 'HEAPF32');
     var localMatrixPtr = copy3x3MatrixToWasm(localMatrix);
     var barePointers = [];
     for (var i = 0; i < childrenShaders.length; i++) {
@@ -35,10 +39,10 @@ CanvasKit._extraInitializations.push(function() {
       // and send that over the wire, so it can be re-wrapped as an sk_sp.
       barePointers.push(childrenShaders[i].$$.ptr);
     }
-    var childrenPointers = copy1dArray(barePointers, "HEAPU32");
+    var childrenPointers = copy1dArray(barePointers, 'HEAPU32');
     // Our array has 4 bytes per float, so be sure to account for that before
     // sending it over the wire.
-    return this._makeShaderWithChildren(fptr, floats.length * 4, childrenPointers,
+    return this._makeShaderWithChildren(fptr, floats.length * 4, shouldOwnUniforms, childrenPointers,
                                         barePointers.length, localMatrixPtr);
   }
 });